arreglo de modulo de avances
