--- SSDAE/app/src/components/avances/CRUDAvance.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const CRUDAvance = () => {
-  return (
-    <div>CRUDAvance</div>
-  )
-}
