+++ SSDAE/app/src/components/avances/EditarAvance.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const EditarAvance = () => {
+  return (
+    <div>EditarAvance</div>
+  )
+}
