+++ SSDAE/app/src/components/avances/NoTesis.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const NoTesis = () => {
+    return (
+        <div className="login-page">
+        <div className='encabezado-avances'>
+            <p>No se tiene asignado ningun tema de tesis</p>
+        </div>
+        </div>
+    )
+}
