+++ SSDAE/app/src/components/avances/SubirAvance.jsx
@@ -1,0 +1,22 @@
+import React, { useRef } from 'react'
+import { useEffect } from 'react';
+import { useState } from 'react';
+
+
+export const SubirAvance = ({setFile, avances, tesis}) => {
+    const saveFile = (e) => {
+        setFile(e.target.files[0]);
+    }
+
+    return (
+        <div>
+            <label htmlFor="file-upload" className="custom-file-upload">
+                Subir archivo
+            </label>
+
+            <input onChange={saveFile} name='file' id="file-upload" type="file" />
+
+            <p style={{ display: 'none' }} ></p>
+        </div>
+    )
+}
