--- SSDAE/app/src/components/avances/texto.txt
@@ -1,2 +1,0 @@
-asdasd
-
