--- SSDAE/app/src/components/texto.txt
@@ -1,2 +1,0 @@
-asdasd
-
