--- SSDAE/app/src/texto.txt
@@ -1,2 +1,0 @@
-asdasd
-
